# [TableGen] Optimize intrinsic info type signature encoding