@jurahul 2024-09-02
Actual benchmark runs:

```
Old:

~/upstream_clean$ ./build/benchmarks/GetIntrinsicInfoTableEntriesBM --benchmark_repetitions=40 --benchmark_display_aggregates_only
2024-09-01T08:16:40-07:00
Running ./build/benchmarks/GetIntrinsicInfoTableEntriesBM
Run on (20 X 2918.4 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x10)
  L1 Instruction 32 KiB (x10)
  L2 Unified 1280 KiB (x10)
  L3 Unified 24576 KiB (x1)
Load Average: 0.43, 0.29, 0.40
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
BM_GetIntrinsicInfoTableEntries_mean       511908 ns       511877 ns           40
BM_GetIntrinsicInfoTableEntries_median     510461 ns       510461 ns           40
BM_GetIntrinsicInfoTableEntries_stddev       9411 ns         9428 ns           40
BM_GetIntrinsicInfoTableEntries_cv           1.84 %          1.84 %            40


new:

~/upstream_llvm/llvm-project$ ./build/benchmarks/GetIntrinsicInfoTableEntriesBM --benchmark_repetitions=40 --benchmark_display_aggregates_only
2024-09-01T08:15:59-07:00
Running ./build/benchmarks/GetIntrinsicInfoTableEntriesBM
Run on (20 X 2918.4 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x10)
  L1 Instruction 32 KiB (x10)
  L2 Unified 1280 KiB (x10)
  L3 Unified 24576 KiB (x1)
Load Average: 0.19, 0.23, 0.39
---------------------------------------------------------------------------------
Benchmark                                       Time             CPU   Iterations
---------------------------------------------------------------------------------
BM_GetIntrinsicInfoTableEntries_mean       474298 ns       474293 ns           40
BM_GetIntrinsicInfoTableEntries_median     474271 ns       474270 ns           40
BM_GetIntrinsicInfoTableEntries_stddev       5285 ns         5287 ns           40
BM_GetIntrinsicInfoTableEntries_cv           1.11 %          1.11 %            40
```

Speedup = 511877  -> 474293  = 7.3%